Update Filament resource structure: webhook module
